requirements.txt
# Core dependencies
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-mcp>=0.1.0
anthropic>=0.40.0

# CLI
click>=8.1.0
rich>=13.0.0

# Data validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# HTTP client
httpx>=0.27.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
diskcache>=5.6.0
tenacity>=8.0.0
markdown>=3.7.0
